[#885 state:resolved] Ported every listing page in Melody to use the jQuery list filter plugin for Melody. This new plugin obviates the need for a lot of hard coded javascript in each of the listing screen's HTML <head>. It also refines how the default listng filter is selected, using an approved setfilter() method on the $.fn.listfilter() plugin.
A number of changes are made to the DOM to support this new plugin because in the OLD code, references to elements were all hard coded. Instead, I needed the plugin to be more context sensitive and aware and be more data driven. Therefore I added filter-value classes to filter value input fields. I added filter-toggle class to which to bind a filter toggle (obviously). Plus a few more odds and ends.